// src/types/IState.ts
export default interface IState {
    priceDomain: number;
    priceDomainCom: number;
    priceDomainComBr: number;
    priceHostingStatic: number;
    priceHostingDynamic: number;
    priceEmail: number;
    priceTotalYearly: number;
    priceTotalMonthly: number;
    disabledFormSubscriptions: boolean;
}